+++ schematics/dependencies.ts
@@ -1,0 +1,5 @@
+export const DEPENDENCIES = {
+    '@angular/cdk': '^9.0.0',
+    'date-fns': '^2.6.0',
+    '@tethys/icons': '*'
+};