+++ schematics/ng-add/index.spec.ts
@@ -1,14 +1,42 @@
 import { Tree } from '@angular-devkit/schematics';
 import { SchematicTestRunner, UnitTestTree } from '@angular-devkit/schematics/testing';
+import { DEPENDENCIES } from '../dependencies';
 import { createTestApp, getJsonFileContent } from '../testing';
+import { addPackageToPackageJson } from '../utils';
+import { VERSION } from '../version';
 
 describe('ng-add Schematic', () => {
     let tree: Tree;
     const schematicRunner = new SchematicTestRunner('ngx-tethys', require.resolve('../collection.json'));
 
     let workspaceTree: UnitTestTree;
+
     beforeEach(async () => {
         tree = await createTestApp(schematicRunner);
+    });
+
+    it('should update package.json', async () => {
+        workspaceTree = await schematicRunner.runSchematicAsync('ng-add', undefined, tree).toPromise();
+        const packageJson = getJsonFileContent(workspaceTree, '/package.json');
+        const dependencies = packageJson.dependencies;
+        expect(dependencies['date-fns']).toEqual(DEPENDENCIES['date-fns']);
+        expect(dependencies['@tethys/icons']).toBeTruthy();
+        expect(dependencies['@angular/cdk']).toEqual(DEPENDENCIES['@angular/cdk']);
+        expect(dependencies['ngx-tethys']).toEqual(VERSION);
+        expect(schematicRunner.tasks.some(task => task.name === 'node-package')).toBe(true);
+    });
+
+    it('should respect version range from CLI ng-add command', async () => {
+        // Simulates the behavior of the CLI `ng add` command. The command inserts the
+        // requested package version into the `package.json` before the actual schematic runs.
+        addPackageToPackageJson(tree, 'ngx-tethys', '^9.0.0');
+
+        workspaceTree = await schematicRunner.runSchematicAsync('ng-add', {}, tree).toPromise();
+        const packageJson = getJsonFileContent(workspaceTree, '/package.json');
+        const dependencies = packageJson.dependencies;
+
+        expect(dependencies['ngx-tethys']).toBe('^9.0.0');
+        expect(schematicRunner.tasks.some(task => task.name === 'node-package')).toBe(true);
     });
 
     it(`should add styles`, async () => {
@@ -22,6 +50,7 @@ describe('ng-add Schematic', () => {
 
         expect(existStyle).toBe(true);
     });
+
     it(`should add styles in [ngx-tethys-chen] project`, async () => {
         tree = await createTestApp(schematicRunner, { name: 'ngx-tethys-chen' }, tree);
         workspaceTree = await schematicRunner.runSchematicAsync('ng-add', { project: 'ngx-tethys-chen' }, tree).toPromise();
@@ -33,6 +62,7 @@ describe('ng-add Schematic', () => {
 
         expect(existStyle).toBe(true);
     });
+
     it(`should add icons`, async () => {
         workspaceTree = await schematicRunner.runSchematicAsync('ng-add', { icon: true }, tree).toPromise();