+++ schematics/testing/index.ts
@@ -1,3 +1,4 @@
 export * from './get-json-content';
-export * from './create-mock-workspace';
 export * from './create-test-environment';
+export * from './test-project';
+export * from './test-app';