+++ src/layout/test/sidebar.spec.ts
@@ -4,7 +4,7 @@ import { ThyLayoutModule } from '../layout.module';
 import { By } from '@angular/platform-browser';
 import { ThyLayoutComponent } from '../layout.component';
 import { ThyHeaderComponent } from '../header.component';
-import { injectDefaultSvgIconSet, bypassSanitizeProvider } from '../../core/testing/thy-icon';
+import { injectDefaultSvgIconSet, bypassSanitizeProvider } from 'ngx-tethys/testing/thy-icon';
 import { ThyContentComponent } from '../content.component';
 import { ThySidebarComponent } from '../sidebar.component';